feat: move generic to module
