+++ libs/@local/hashql/core/src/type/kind/generic/mod.rs
@@ -1,7 +1,10 @@
+pub mod apply;
+
 use core::ops::Deref;
 
 use pretty::RcDoc;
 
+pub use self::apply::{Apply, GenericSubstitution, GenericSubstitutions};
 use crate::{
     intern::Interned,
     newtype, newtype_producer,
@@ -136,139 +139,6 @@ impl PrettyPrint for GenericArguments<'_> {
                 )
                 .append(RcDoc::text(">")),
         }
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct GenericSubstitution {
-    pub argument: GenericArgumentId,
-    pub value: TypeId,
-}
-
-impl PrettyPrint for GenericSubstitution {
-    fn pretty<'env>(
-        &self,
-        env: &'env Environment,
-        limit: RecursionDepthBoundary,
-    ) -> RcDoc<'env, anstyle::Style> {
-        let name = format!("?{}", self.argument);
-
-        RcDoc::text(name)
-            .annotate(ORANGE)
-            .append(RcDoc::line())
-            .append("=")
-            .append(RcDoc::line())
-            .append(limit.pretty(env, self.value))
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct GenericSubstitutions<'heap>(Option<Interned<'heap, [GenericSubstitution]>>);
-
-impl<'heap> GenericSubstitutions<'heap> {
-    #[must_use]
-    pub const fn empty() -> Self {
-        Self(None)
-    }
-
-    /// Create a new `GenericSubstitutions` from a slice of `GenericSubstitution`s.
-    ///
-    /// The caller must ensure that the slice is sorted by argument ID and contains no duplicates.
-    ///
-    /// You should probably use `Environment::intern_generic_substitutions` instead.
-    #[must_use]
-    pub const fn from_slice_unchecked(slice: Interned<'heap, [GenericSubstitution]>) -> Self {
-        Self(Some(slice))
-    }
-
-    #[must_use]
-    pub const fn as_slice(&self) -> &[GenericSubstitution] {
-        match self.0 {
-            Some(Interned(slice, _)) => slice,
-            None => &[] as &[GenericSubstitution],
-        }
-    }
-
-    #[must_use]
-    pub const fn len(&self) -> usize {
-        self.as_slice().len()
-    }
-
-    #[must_use]
-    pub const fn is_empty(&self) -> bool {
-        self.as_slice().is_empty()
-    }
-
-    #[must_use]
-    pub fn merge(&self, other: &Self, env: &Environment<'heap>) -> Self {
-        // We can merge without de-duplication, because every argument has a unique ID.
-        // What we need to do tho, is to re-sort them, so that the invariants are maintained.
-        let mut vec = Vec::with_capacity(self.len() + other.len());
-
-        vec.extend_from_slice(self.as_slice());
-        vec.extend_from_slice(other.as_slice());
-
-        env.intern_generic_substitutions(&mut vec)
-    }
-}
-
-impl AsRef<[GenericSubstitution]> for GenericSubstitutions<'_> {
-    fn as_ref(&self) -> &[GenericSubstitution] {
-        self.as_slice()
-    }
-}
-
-impl Deref for GenericSubstitutions<'_> {
-    type Target = [GenericSubstitution];
-
-    fn deref(&self) -> &Self::Target {
-        self.as_slice()
-    }
-}
-
-impl PrettyPrint for GenericSubstitutions<'_> {
-    fn pretty<'env>(
-        &self,
-        env: &'env Environment,
-        limit: RecursionDepthBoundary,
-    ) -> RcDoc<'env, anstyle::Style> {
-        match self.as_slice() {
-            [] => RcDoc::nil(),
-            slice => RcDoc::text("<")
-                .append(
-                    RcDoc::intersperse(
-                        slice
-                            .iter()
-                            .map(|substitution| substitution.pretty(env, limit)),
-                        RcDoc::text(",").append(RcDoc::line()),
-                    )
-                    .nest(1)
-                    .group(),
-                )
-                .append(RcDoc::text(">")),
-        }
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct Apply<'heap> {
-    pub base: TypeId,
-    pub substitutions: GenericSubstitutions<'heap>,
-}
-
-impl PrettyPrint for Apply<'_> {
-    fn pretty<'env>(
-        &self,
-        env: &'env Environment,
-        limit: RecursionDepthBoundary,
-    ) -> RcDoc<'env, anstyle::Style> {
-        limit.pretty(env, self.base).append(
-            RcDoc::line()
-                .append(RcDoc::text("where").annotate(RED))
-                .append(self.substitutions.pretty(env, limit))
-                .group()
-                .nest(1),
-        )
     }
 }
